old-tests/submission/Microsoft/table/table_001.htm

<!DOCTYPE HTML>
<html>
<head>
<title>HTML5 Table API Tests</title>
  <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
  <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
  <link rel="help" href="http://www.w3.org/TR/html5/tabular-data.html#tabular-data" />
</head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<table id="table1" style="display:none">
<tr><td></td></tr>
<caption>first caption</caption>
<caption>second caption</caption>
</table>
<table id="table2" style="display:none">
<tr><td></td></tr>
</table>
<table id="table3" style="display:none">
<tr><td><caption>non-child caption</caption></td></tr>
<caption>first caption</caption>
<caption>second caption</caption>
</table>
<table id="table4" style="display:none">
<tr><td></td></tr>
<caption>first caption</caption>
</table>
<script>
    test(function () {
        assert_equals(document.getElementById('table1').caption.innerHTML, "first caption");
    }, "first caption element child of the first table element")

    function captionElement() {
        var caption = document.createElement("caption");
        caption.innerHTML = "new caption";
        document.getElementById('table1').caption = caption;

        assert_equals(document.getElementById('table1').caption.innerHTML, "new caption");
    }
    test(captionElement, "setting caption on a table");

    test(function () {
        assert_equals(document.getElementById('table2').caption, null);
    }, "caption IDL attribute is null")

    test(function () {
        assert_equals(document.getElementById('table3').caption.innerHTML, "non-child caption");
    }, "first caption element child of the third table element")

    function removeCaptionElement() {
        var parent = document.getElementById('table4').caption.parentNode;
        parent.removeChild(document.getElementById('table4').caption);

        assert_equals(document.getElementById('table4').caption, null);
    }
    test(captionElement, "dynamically removing caption on a table");
</script>
</body>
</html>

